Cut down on some of the serialised byte data size for a compiled script/expression.

+++ DialogML/VM/ScriptEngine.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 
 namespace DialogML
 {
-    enum XNodeType : ushort
+    enum XNodeType : byte
     {
         Unknown = 0,
         Script = 1,
